day15/autoquant-alpha-d15/requirements.txt
alpaca-py>=0.20.0
numpy>=1.26.0
pandas>=2.2.0
rich>=13.7.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
